move more ui.js functions to according modules
