+++ js/ui.js
@@ -1,4 +1,4 @@
-/* global Notification, adjustControlsWidth, SearchProxy */
+/* global adjustControlsWidth */
 
 /**
  * ownCloud - Mail
@@ -15,11 +15,8 @@ define(function(require) {
 
 	var $ = require('jquery');
 	var Handlebars = require('handlebars');
-	var Marionette = require('marionette');
 	var OC = require('OC');
 	var Radio = require('radio');
-	var MessagesView = require('views/messages');
-	var NavigationAccountsView = require('views/navigation-accounts');
 	var ComposerView = require('views/composer');
 
 	require('views/helper');
@@ -32,9 +29,7 @@ define(function(require) {
 	});
 	Radio.ui.on('composer:leave', onComposerLeave);
 	Radio.ui.on('folder:changed', onFolderChanged);
-	Radio.ui.on('window:resize', onWindowResize);
 
-	var messageView = null;
 	var composer = null;
 	var composerVisible = false;
 
@@ -62,116 +57,13 @@ define(function(require) {
 		$('#mail-message').addClass('hidden-mobile');
 	}
 
-	function onWindowResize() {
-		// Resize iframe
-		var iframe = $('#mail-content iframe');
-		iframe.height(iframe.contents().find('html').height() + 20);
-
-		// resize width of attached images
-		$('.mail-message-attachments .mail-attached-image').each(function() {
-			$(this).css('max-width', $('.mail-message-body').width());
-		});
-	}
-
 	function initializeInterface() {
-		// Register UI events
-		window.addEventListener('resize', Radio.ui.on('window:resize'));
-
-		Marionette.TemplateCache.prototype.compileTemplate = function(rawTemplate) {
-			return Handlebars.compile(rawTemplate);
-		};
-		Marionette.ItemView.prototype.modelEvents = {'change': 'render'};
-		Marionette.CompositeView.prototype.modelEvents = {'change': 'render'};
-
-		// ask to handle all mailto: links
-		if (window.navigator.registerProtocolHandler) {
-			var url = window.location.protocol + '//' +
-				window.location.host +
-				OC.generateUrl('apps/mail/compose?uri=%s');
-			try {
-				window.navigator
-					.registerProtocolHandler('mailto', url, 'ownCloud Mail');
-			} catch (e) {
-			}
-		}
-
-		// setup messages view
-		messageView = new MessagesView({
-			el: $('#mail-messages')
-		});
-		messageView.render();
-
-		// setup folder view
-		var foldersView = new NavigationAccountsView();
-		require('state').folderView = foldersView;
-		require('app').view.navigation.accounts.show(foldersView);
-
-		foldersView.listenTo(messageView, 'change:unseen',
-			foldersView.changeUnseen);
-
-		// request permissions
-		if (typeof Notification !== 'undefined') {
-			Notification.requestPermission();
-		}
-
-		// Register actual filter method
-		SearchProxy.setFilter(require('search').filter);
-
-		function split(val) {
-			return val.split(/,\s*/);
-		}
-
-		function extractLast(term) {
-			return split(term).pop();
-		}
-		$(document).on('focus', '.recipient-autocomplete', function() {
-			if (!$(this).data('autocomplete')) { // If the autocomplete wasn't called yet:
-				// don't navigate away from the field on tab when selecting an item
-				$(this).bind('keydown', function(event) {
-					if (event.keyCode === $.ui.keyCode.TAB &&
-						typeof $(this).data('autocomplete') !== 'undefined' &&
-						$(this).data('autocomplete').menu.active) {
-						event.preventDefault();
-					}
-				}).autocomplete({
-					source: function(request, response) {
-						$.getJSON(
-							OC.generateUrl('/apps/mail/autoComplete'),
-							{
-								term: extractLast(request.term)
-							}, response);
-					},
-					search: function() {
-						// custom minLength
-						var term = extractLast(this.value);
-						return term.length >= 2;
-
-					},
-					focus: function() {
-						// prevent value inserted on focus
-						return false;
-					},
-					select: function(event, ui) {
-						var terms = split(this.value);
-						// remove the current input
-						terms.pop();
-						// add the selected item
-						terms.push(ui.item.value);
-						// add placeholder to get the comma-and-space at the end
-						terms.push('');
-						this.value = terms.join(', ');
-						return false;
-					}
-				});
-			}
-		});
-
 		setInterval(require('background').checkForNotifications, 5 * 60 * 1000);
 		Radio.account.trigger('load');
 	}
 
 	function clearMessages() {
-		messageView.collection.reset();
+		Radio.ui.trigger('messagesview:messages:reset');
 		$('#messages-loading').fadeIn();
 
 		$('#mail-message')
@@ -185,7 +77,7 @@ define(function(require) {
 	}
 
 	function addMessages(data) {
-		messageView.collection.add(data);
+		Radio.ui.trigger('messagesview:messages:add', data);
 	}
 
 	function loadFolder(accountId, folderId, noSelect) {
@@ -265,7 +157,7 @@ define(function(require) {
 					if (cached) {
 						// Trigger folder update
 						// TODO: replace with horde sync once it's implemented
-						messageView.loadNew();
+						Radio.ui.trigger('messagesview:messages:update');
 					}
 				},
 				onError: function(error, textStatus) {
@@ -551,7 +443,7 @@ define(function(require) {
 				});
 			});
 
-			messageView.setMessageFlag(messageId, 'unseen', false);
+			Radio.ui.trigger('messagesview:messageflag:set', messageId, 'unseen', false);
 
 			// HTML mail rendering
 			$('iframe').load(function() {
@@ -641,7 +533,7 @@ define(function(require) {
 	}
 
 	function setFolderActive(accountId, folderId) {
-		messageView.clearFilter();
+		Radio.ui.trigger('messagesview:filter:clear');
 
 		// disable all other folders for all accounts
 		require('state').accounts.each(function(account) {
@@ -660,7 +552,7 @@ define(function(require) {
 	}
 
 	function setMessageActive(messageId) {
-		messageView.setActiveMessage(messageId);
+		Radio.ui.trigger('messagesview:message:setactive', messageId);
 		require('state').currentMessageId = messageId;
 		require('state').folderView.updateTitle();
 	}
@@ -681,17 +573,6 @@ define(function(require) {
 		setFolderActive: setFolderActive,
 		setMessageActive: setMessageActive
 	};
-
-	Object.defineProperties(view, {
-		messageView: {
-			get: function() {
-				return messageView;
-			},
-			set: function(mv) {
-				messageView = mv;
-			}
-		}
-	});
 
 	return view;
 });
